chore(assets): rename calendar illustration file
