+++ edgar-app/src/assets/illustrations/CalendarIllustration.tsx
